requirements.txt
click~=8.1.7
PyYAML~=6.0.2
protobuf~=5.28.3
google-auth-oauthlib~=1.2.1
google-api-python-client~=2.154.0